% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_asym.R
\name{asym_function}
\alias{asym_function}
\title{Asymptotic function for the optimize function}
\usage{
asym_function(estimated, X, Y)
}
\arguments{
\item{estimated}{A vector of the current guesses on the coefficients of the
model.}

\item{X}{A matrix of the predictor space}

\item{Y}{A vector of the observed results used to calculate the Loss
function.}
}
\value{
A numeric value representing the error with the current parameter
  estimates.
}
\description{
It returns the loss of the additive sinusoidal function for the
  [stats::optim()] function.
}
