% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_asym.R
\name{asym_optimize}
\alias{asym_optimize}
\title{Optim function for the Additive Asymptotic Regression Model}
\usage{
asym_optimize(init_guess, X, Y, method, ...)
}
\arguments{
\item{init_guess}{The initial parameter guesses for the optim function, by
default it is all ones.}

\item{X}{A matrix of the predictor space}

\item{Y}{A vector of the observed results used to calculate the Loss
function.}

\item{method}{The method to be used. See method in
\code{\link[stats]{optim}}.}

\item{...}{Additional arguments passed to the \code{\link[stats]{optim}}
function.}
}
\value{
For \code{optim}, a list with components:
  \item{par}{The best set of parameters found.}
  \item{value}{The value of \code{fn} corresponding to \code{par}.}
  \item{counts}{A two-element integer vector giving the number of calls
    to \code{fn} and \code{gr} respectively. This excludes those calls needed
    to compute the Hessian, if requested, and any calls to \code{fn} to
    compute a finite-difference approximation to the gradient.}
  \item{convergence}{An integer code. \code{0} indicates successful
    completion (which is always the case for \code{"SANN"} and
    \code{"Brent"}).  Possible error codes are
    \describe{
      \item{\code{1}}{indicates that the iteration limit \code{maxit}
      had been reached.}
      \item{\code{10}}{indicates degeneracy of the Nelder--Mead simplex.}
      \item{\code{51}}{indicates a warning from the \code{"L-BFGS-B"}
      method; see component \code{message} for further details.}
      \item{\code{52}}{indicates an error from the \code{"L-BFGS-B"}
      method; see component \code{message} for further details.}
    }
  }
  \item{message}{A character string giving any additional information
    returned by the optimizer, or \code{NULL}.}
  \item{hessian}{Only if argument \code{hessian} is true. A symmetric
    matrix giving an estimate of the Hessian at the solution found.  Note
    that this is the Hessian of the unconstrained problem even if the
    box constraints are active.}

  For \code{optimHess}, the description of the \code{hessian} component
  applies.
}
\description{
Utilizes the [stats::optim()] function to perform the
  optimization of the curve.
}
