% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{boolean_check}
\alias{boolean_check}
\title{Boolean Check function}
\usage{
boolean_check(boolean)
}
\arguments{
\item{boolean}{A Boolean value}
}
\value{
Allows the function to run or returns a user friendly error message.
}
\description{
A helper function to check if value is a Boolean value.
}
