% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_part.R
\name{cluster_part}
\alias{cluster_part}
\title{Cluster-based partitions for cross validation}
\usage{
cluster_part(features, k, k_mult = 5, ...)
}
\arguments{
\item{features}{A scaled matrix of features to be used in the clustering.
Scaling usually done with [base::scale()] and should not include the
predictor variable.}

\item{k}{The number of partitions for k-fold cross validation.}

\item{k_mult}{k*k_mult determines the number of subgroups that will be
created as part of the balancing algorithm.}

\item{...}{Additional arguments passed to [stats::kmeans()] as needed.}
}
\description{
This function creates cluster-based partitions of a sample space based on
  k-means clustering. Included in the function are algorithms that attempt
  to produce clusters of roughly equal size.
}
