% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_part2.R
\name{cluster_part2}
\alias{cluster_part2}
\title{Spatial cluster-based partitions for cross validation}
\usage{
cluster_part2(features, k, k_mult = 5, ...)
}
\arguments{
\item{features}{A scaled matrix of features to be used in the clustering.
Scaling usually done with [base::scale()] and should not include the
predictor variable.}

\item{k}{The number of partitions for k-fold cross validation.}

\item{...}{Additional arguments passed to [stats::kmeans()] as needed.}

\item{k2}{If specified, k2 represents a finer scale partitioning that can be
used the balance the larger k-fold clusters. k2 must be at least double
k if it is to be used effectively (the function will throw an error if
it is not).}
}
\description{
This function creates cluster-based partitions of a sample space based on
  k-means clustering. Included in the function are algorithms that attempt
  to produce clusters of roughly equal size.
}
