% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mlm.R
\name{compare_mlm}
\alias{compare_mlm}
\title{Compare Machine Learning Models}
\usage{
compare_mlm(formula, data, n_models = 9999, classification = FALSE, ...)
}
\arguments{
\item{formula}{The regression formula or classification formula}

\item{data}{A dataframe object that includes the test data}

\item{n_models}{A defaulted integer to return the maximum number of models}

\item{classification}{A boolean value tag to indicate if classification
methods should be used.}

\item{...}{Additional arguments passed to the setup function in pycaret}
}
\value{
A list object that contains all the fitted models and the CV
  predictive accuracy.
}
\description{
Through the pycaret module from python, this function fits many machine
  learning models simultaneously with little to no coding on the user
  part.
}
