% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lm.R
\name{cv.lm}
\alias{cv.lm}
\title{Cross Validation for lm}
\usage{
\method{cv}{lm}(object, data, n_folds = 10, k_mult = NULL)
}
\arguments{
\item{object}{A lm object.}

\item{data}{A data-frame object to be used for cross-validation}

\item{n_folds}{An integer value for the number of folds defaulted to 10. If
NULL, it will run LOO cross validation.}

\item{k_mult}{Used to specify if k-means clustering is to be used, defaulted
to NULL.}
}
\value{
A vector of cross-validated predictions.
}
\description{
This is the cross validation wrapper for the lm method.
}
\examples{
 lm_test <- data_gen_lm(10)
 lm <- lm(Y ~ ., lm_test)
 cv(lm, lm_test, n_folds = 5)
}
