% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mlm_stressor.R
\name{cv.mlm_stressor}
\alias{cv.mlm_stressor}
\title{Cross Validation for mlm_stressor}
\usage{
\method{cv}{mlm_stressor}(object, data, n_folds = 10, k_mult = NULL)
}
\arguments{
\item{object}{A reg_sine object.}

\item{data}{A data-frame object to be used for cross-validation}

\item{n_folds}{An integer value for the number of folds defaulted to 10. If
NULL, it will run LOO cross validation.}

\item{k_mult}{Used to specify if k-means clustering is to be used, defaulted
to NULL.}
}
\value{
A data frame of cross-validated predictions where the columns are the
  various methods of machine learning models.
}
\description{
This is the cross validation wrapper for the mlm_stressor method
}
