% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_reg_asym.R
\name{cv.reg_asym}
\alias{cv.reg_asym}
\title{Cross Validation for reg_asym}
\usage{
\method{cv}{reg_asym}(object, data, n_folds = 10, k_mult = NULL)
}
\arguments{
\item{object}{A reg_asym object.}

\item{data}{A data-frame object to be used for cross-validation}

\item{n_folds}{An integer value for the number of folds defaulted to 10.}

\item{k_mult}{Used to specify if k-means clustering is to be used, defaulted
to NULL.}
}
\value{
A vector of cross-validated predictions.
}
\description{
This is the cross validation wrapper for the reg_asym method.
}
