% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_reg_sine.R
\name{cv.reg_sine}
\alias{cv.reg_sine}
\title{Cross Validation for reg_sine}
\usage{
\method{cv}{reg_sine}(object, data, n_folds = 10, k_mult = NULL)
}
\arguments{
\item{object}{A reg_sine object.}

\item{data}{A data-frame object to be used for cross-validation}

\item{n_folds}{An integer value for the number of folds defaulted to 10. If
NULL, it will run LOO cross validation.}

\item{k_mult}{Used to specify if k-means clustering is to be used, defaulted
to NULL.}
}
\value{
A vector of cross-validated predictions.
}
\description{
This is the cross validation wrapper for the reg_sine method.
}
