% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{data_check}
\alias{data_check}
\title{Data and Formula helper check function}
\usage{
data_check(formula, data, train_data = NULL)
}
\arguments{
\item{formula}{A formula object}

\item{data}{A data.frame object}

\item{train_data}{A data.frame object. Some methods include two data.frames
as inputs.}
}
\value{
Stops the function from continuing it's output by displaying an
 error message or nothing is returned.
}
\description{
Checks that the formula is a formula object, data is data.frame object, and
 if all the variables in the formula call are in the data.frame
}
