% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{integer_check}
\alias{integer_check}
\title{Integer Check function}
\usage{
integer_check(integer)
}
\arguments{
\item{integer}{An integer value}
}
\value{
Allows the function to run or returns a user friendly error message.
}
\description{
A helper function to check if value is an integer
}
