% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_classification.R
\name{kappa_class}
\alias{kappa_class}
\title{Kappa function}
\usage{
kappa_class(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{A matrix or table that is the confusion matrix.}
}
\value{
A numeric value representing the kappa value.
}
\description{
A function to calculate the Kappa of binary classification.
}
