% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{matrix_check}
\alias{matrix_check}
\title{Matrix Check function}
\usage{
matrix_check(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\value{
Allows the function to run or returns a user friendly error message.
}
\description{
A helper function to check if the matrix is properly formatted
}
