% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_thinning.R
\name{mlm_thinning}
\alias{mlm_thinning}
\title{Thinning algorithim for Machine Learning Models}
\usage{
mlm_thinning(mlm_object, data, max = 0.95, min = 0.05, iter = 0.05)
}
\arguments{
\item{mlm_object}{a `"mlm_stressor"` object that has models in the first list
object.}

\item{data}{a data frame with all the data}

\item{max}{a numeric value in (0, 1] and greater than `min`,
defaulted to .95.}

\item{min}{a numeric value in (0, 1) and less than `max`, defaulted to .05.}

\item{iter}{a numeric value to indicate the step size, defaulted to .05}
}
\value{
a list object where each item in the list is one step in the thinning
 process and is a data frame of predictions.
}
\description{
Fits various train size and test sizes.
}
\examples{
 lm_test <- data_gen_lm(10)
 create_virtualenv()
 mlm_lm <- mlm_regressor(Y ~ ., lm_test, example = TRUE)
 thin <- mlm_thinning(mlm_lm, lm_test, max = .8, min = .7, iter = .05)
 thin
}
