% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlm_vignette_boston}
\alias{mlm_vignette_boston}
\title{Machine Learning Models fitted to Boston data}
\format{
\describe{
  \item{models}{List of the models currently supported in PyCaret}
  \item{pred_accuracy}{The initial predictive accuracy of the models}
}
}
\usage{
mlm_vignette_boston
}
\description{
Fitted to the Boston data set in this package to save computational time for
 the vignette.
}
\keyword{datasets}
