% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlm_vignette_boston_cv}
\alias{mlm_vignette_boston_cv}
\title{Machine Learning Models fitted to Boston data Cross Validated}
\format{
A data frame with 18 columns and 506 rows where each column represents a
 different machine learning model.
\describe{
  \item{ada}{AdaBoost Regressor}
  \item{br}{Bayesian Ridge}
  \item{dt}{Decision Tree Regressor}
  \item{dummy}{Dummy Regressor}
  \item{en}{Elastic Net}
  \item{et}{Extra Trees Regressor}
  \item{gbr}{Gradient Boosting Regressor}
  \item{huber}{Huber Regressor}
  \item{knn}{K Neighbors Regressor}
  \item{lar}{Least Angle Regression}
  \item{lasso}{Lasso Regression}
  \item{lightgbm}{Light Gradient Boosting Machine}
  \item{llar}{Lasso Least Angle Regression}
  \item{lr}{Linear Regression}
  \item{omp}{Orthogonal Matching Pursuit}
  \item{par}{Passive Aggressive Regressor}
  \item{rf}{Random Forest Regressor}
  \item{ridge}{Ridge Regression}
}
}
\usage{
data(mlm_vignette_boston_cv)
}
\description{
Fitted to the Boston data set in this package to save
 computational time for the vignette.
}
\keyword{datasets}
