% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlm_vignette_boston_pred}
\alias{mlm_vignette_boston_pred}
\title{Test predictive accuracy of Machine Learning Models on Boston data}
\format{
A data frame with 2 columns and 18 rows where each row is a different
   machine learning model.
\describe{
 \item{models}{Machine Learning Models}
 \item{rmse}{Root mean square error}
}
}
\usage{
data(mlm_vignette_boston_pred)
}
\description{
Fitted to the Boston data set with the results from the
  predictive accuracy data frame returned.
}
\keyword{datasets}
