% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlm_vignette_lm}
\alias{mlm_vignette_lm}
\title{Machine Learning Models fitted to Boston data}
\format{
\describe{
  \item{models}{List of the models currently supported in PyCaret}
  \item{pred_accuracy}{The initial predictive accuracy of the models}
}
}
\usage{
mlm_vignette_lm
}
\description{
Fitted to a linear data set in this package to save computational time for
 the vignette. This data set was created by `"data_gen_lm()"` function in
 this package
}
\keyword{datasets}
