% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{numeric_check}
\alias{numeric_check}
\title{Numeric Check function}
\usage{
numeric_check(numeric)
}
\arguments{
\item{numeric}{A numeric value}
}
\value{
Allows the function to run or returns a user friendly error message.
}
\description{
A helper function to check if value is an numeric
}
