% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mlm_stressor.R
\name{predict.mlm_stressor}
\alias{predict.mlm_stressor}
\title{Predict for Machine Learning Models Object}
\usage{
\method{predict}{mlm_stressor}(object, newdata, train_data = NULL, ...)
}
\arguments{
\item{object}{A `"mlm_stressor"` object}

\item{newdata}{A data.frame object that is to be predicted}

\item{train_data}{Defaulted to `NULL`, but a data.frame when specified for
the new training to take place.}

\item{...}{Additional arguments that have to be overwritten currently not be
passed on.}
}
\value{
A data.frame object where the columns are the various methods and the
 rows are the predictions
}
\description{
A prediction method to fit predictions given test data as well as can refit
 and predict if given training data to refit on.
}
\examples{
 lm_test <- data_gen_lm(10)
 create_virtualenv()
 mlm_lm <- mlm_regressor(Y ~ ., lm_test, example = TRUE)
 predict(mlm_lm, lm_test)
}
