% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_reg_asym.R
\name{predict.reg_asym}
\alias{predict.reg_asym}
\title{Predict for Asymptotic Regression}
\usage{
\method{predict}{reg_asym}(object, newdata, ...)
}
\arguments{
\item{object}{A "reg_asym" object that has a formula attribute and parameter
estimates from \link[reg_asym]{reg_asym}.}

\item{newdata}{Dataframe that has the same variables as the formula from
"reg_asym".}
}
\value{
A vector with the predicted values.
}
\description{
Predict values on an additive exponential model
}
