% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_reg_sine.R
\name{predict.reg_sine}
\alias{predict.reg_sine}
\title{Predict for Sinusoidal Regression}
\usage{
\method{predict}{reg_sine}(object, newdata, ...)
}
\arguments{
\item{object}{A "reg_sine" object that has a formula attribute and parameter
estimates from \link[reg_sine]{reg_sine}.}

\item{newdata}{Dataframe that has the same variables as the formula from
"reg_sine".}

\item{...}{Extending the [stats::predict()] function default, in this case
this is ignored.}
}
\value{
A vector with the predicted values.
}
\description{
Predict values on an additive exponential model
}
