% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{python_avail}
\alias{python_avail}
\title{Check if `Python` is Available}
\usage{
python_avail()
}
\value{
A Boolean value is returned.
}
\description{
A function that allows examples to run when appropriate.
}
\examples{
 python_avail()
}
