% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{residual}
\alias{residual}
\title{Residual Value Calculation}
\usage{
residual(predictions, observed)
}
\arguments{
\item{predictions}{A vector or data frame of predicted values}

\item{observed}{A vector of the observed values}
}
\value{
A data frame or a vector depending on the input of the predictions
 parameter
}
\description{
Helper function when wanting to calculate the residual values for the
 machine learning models as they return a data frame of predictions. This
 function also handles vectors as well.
}
\examples{
 pred <- matrix(rnorm(50), ncol = 5, nrow = 10)
 obs <- rnorm(10)
 res <- residual(pred, obs)
}
