% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(predictions, observed)
}
\arguments{
\item{predictions}{A data frame or vector object that is the same number of
rows or length as the length of observed values.}

\item{observed}{A vector of the observed results}
}
\value{
A finite value or a data frame of methods and their rmse values
}
\description{
Calculates the root mean square error.
}
\examples{
  lm_data <- data_gen_lm(25)
  test_index <- sample(1:nrow(lm_data), 5)
  test <- lm_data[test_index, ]
  train <- lm_data[-test_index, ]
  lm_test <- lm(Y ~ ., train)
  lm_cv <- cv(lm_test, lm_data, n_folds = 5)
  lm_cv
}
