% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_regression.R
\name{score_regression}
\alias{score_regression}
\title{Score Function for Regression}
\usage{
score_regression(
  observed,
  predicted,
  metrics = c("RMSE", "MAE", "MSE", "R2", "RMSLE", "MAPE")
)
}
\arguments{
\item{observed}{A vector of the observed results.}

\item{predicted}{A data.frame or vector object that is the same number of
rows or length as the length of observed values.}

\item{metrics}{A character vector of the metrics to be fitted. This is
defaulted to be the metrics from `PyCaret`.}
}
\value{
A vector or data.frame of the methods and metrics.
}
\description{
This function takes the observed and predicted values and
 computes metrics that are found in `PyCaret` such as: `RMSE`, `MAE`, `MSE`,
 `R2`, `RMSLE`, and `MAPE`.
}
