% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_sine.R
\name{sine_gradient}
\alias{sine_gradient}
\title{Gradient Sine Function used for Optim}
\usage{
sine_gradient(estimated, X, Y)
}
\arguments{
\item{estimated}{A vector of the current guesses on the coefficients of the
model.}

\item{X}{A matrix of the predictor space}

\item{Y}{A vector of the observed results used to calculate the Loss
function.}
}
\value{
The gradient of the loss with the current parameter estimates
}
\description{
This is the gradient function used for the [stats::optim()] for
 the `"BFGS"` optimization of optim.
}
