% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_sine.R
\name{sine_yhat}
\alias{sine_yhat}
\title{Predictions for the Additive Sinusoidal Model}
\usage{
sine_yhat(estimated, X)
}
\arguments{
\item{estimated}{A vector of the current guesses on the coefficients of the
model.}

\item{X}{A matrix of the predictor space}
}
\value{
A vector of predictions from the model
}
\description{
Fits the Additive Sinusoidal model with the current coefficients
  and the current predictor space.
}
