% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vector_check}
\alias{vector_check}
\title{Vector Check function}
\usage{
vector_check(vector)
}
\arguments{
\item{vector}{A vector}
}
\value{
Allows the function to run or returns a user friendly error message.
}
\description{
A helper function to check if the vector is properly formatted
}
