% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mlm.R
\name{compare_mlm}
\alias{compare_mlm}
\title{Compare Machine Learning Models}
\usage{
compare_mlm(
  formula,
  data,
  n_models = 9999,
  classification = FALSE,
  fit_models = c("ada", "et", "lightgbm", "gbr", "lr", "rf", "ridge", "knn", "dt",
    "dummy", "lar", "br", "huber", "omp", "lasso", "en", "llar", "par"),
  ...
)
}
\arguments{
\item{formula}{The regression formula or classification formula}

\item{data}{A dataframe object that includes the test data}

\item{n_models}{A defaulted integer to return the maximum number of models}

\item{classification}{A boolean value tag to indicate if classification
methods should be used. If set to TRUE, these models are used instead of
fit_models:
\tabular{rl}{
  ada \tab AdaBoost Classifier \cr
  et \tab Extra Trees Classifier \cr
  lightgbm \tab Light Gradient Boosting Machine\cr
  gbr \tab Gradient Boosting Classifier\cr
  lr \tab Logistic Regression\cr
  rf \tab Random Forest Classifier\cr
  ridge \tab Ridge Classifier\cr
  knn \tab K Neighbors Classifier\cr
  dt \tab Decision Tree Classifier\cr
  dummy \tab Dummy Classifier\cr
  svm \tab SVM - Linear Kernel\cr
  lda \tab Linear Discriminant Analysis\cr
  nb \tab Naive Bayes \cr
  qda \tab Quadratic Discriminant Analysis
}}

\item{fit_models}{A character vector with all the possible Machine Learning
regressors
\tabular{rl}{
  ada \tab AdaBoost Regressor \cr
  et \tab Extra Trees Regressor \cr
  lightgbm \tab Light Gradient Boosting Machine\cr
  gbr \tab Gradient Boosting Regressor\cr
  lr \tab Linear Regression\cr
  rf \tab Random Forest Regressor\cr
  ridge \tab Ridge Regression\cr
  knn \tab K Neighbors Regressor\cr
  dt \tab Decision Tree Regressor\cr
  dummy \tab Dummy Regressor\cr
  lar \tab Least Angle Regression\cr
  br \tab Bayesian Ridge\cr
  huber \tab Huber Regressor\cr
  omp \tab Orthogonal Matching Pursuit\cr
  lasso \tab Lasso Regression\cr
  en \tab Elastic Net\cr
  llar \tab Lasso Least Angle Regression\cr
  par \tab Passive Aggressive Regressor
}}

\item{...}{Additional arguments passed to the setup function in pycaret}
}
\value{
A list object that contains all the fitted models and the CV
  predictive accuracy.
}
\description{
Through the pycaret module from python, this function fits many machine
  learning models simultaneously with little to no coding on the user
  part.
}
