% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_groups.R
\name{create_groups}
\alias{create_groups}
\title{Create groups for CV}
\usage{
create_groups(formula, data, n_folds = 10, k_mult = NULL)
}
\arguments{
\item{formula}{A formula object that fits the data}

\item{data}{The data that will be separated into each group}

\item{n_folds}{An integer value defaulted to 10 fold cross-validation if NULL
uses LOO instead.}

\item{k_mult}{When specified is passed onto the \code{\link{cluster_part2}}
to fit the data into k_groups}
}
\value{
A vector of the assigned groups
}
\description{
Creates groups for the data by separating them either into 10 fold cross-
  validation, LOO cross-validation, or k-means grouping.
}
