% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtualenv.R
\name{create_virtualenv}
\alias{create_virtualenv}
\title{Create Python Virtual Environment}
\usage{
create_virtualenv(python = Sys.which("python"), delete_env = FALSE)
}
\arguments{
\item{python}{Defaults to your install of python, prefer 3.8.10}

\item{delete_env}{Boolean value to indicate if the environments need to be
deleted.}
}
\value{
A message indicating which environment is being used
}
\description{
Allows the user to create a stressor python environment with pycaret
  installed on the environment. This function takes into account that
  you have properly installed python. We recommend version 3.8.10. Uses
  existing stressor environments.
}
