% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtualenv.R
\name{create_virtualenv}
\alias{create_virtualenv}
\title{Create Python Virtual Environment}
\usage{
create_virtualenv(python = Sys.which("python"), delete_env = FALSE)
}
\arguments{
\item{python}{Defaults to your install of python, prefer 3.8.10. This is
assuming that you installed python from python.org. Currently Anaconda
installations of python are not implemented.}

\item{delete_env}{Boolean value to indicate if the environments need to be
deleted.}
}
\value{
A message indicating which environment is being used.
}
\description{
Allows the user to create a stressor python environment with PyCaret
  installed on the environment. This function assumes that
  you have properly installed python. We recommend version 3.8.10. Uses
  existing stressor environments.
}
\details{
To install python, It is recommended using Python version 3.8.10 from
 \href{https://www.python.org/downloads/release/python-3810/}{python.org}, as
 this is the same version recommended by PyCaret as it is the most stable.
 Users have reported troubles using the Anaconda distribution of python.
}
\examples{
 create_virtualenv()
}
