% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Cross Validation}
\usage{
cv(object, data, n_folds = 10, k_mult = NULL)
}
\arguments{
\item{object}{One of the four objects that is accepted mlm_stressor,
reg_sine, reg_asym, or lm.}

\item{data}{A data frame object that contains all the entries to be cross-
validated on.}

\item{n_folds}{An integer value for the number of folds defaulted to 10. If
NULL, it will run LOO cross validation.}

\item{k_mult}{Used to specify if k-means clustering is to be used, defaulted
to NULL.}
}
\value{
If the object is of class mlm_stressor then a data frame will be
  returned otherwise a vector of the predictions will be returned.
}
\description{
This is core of cross validation both standard and using k-mean groups. This
 method is called by other cv methods of classes.
}
