% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Cross Validation}
\usage{
cv(object, data, t_groups)
}
\arguments{
\item{object}{One of the four objects that is accepted mlm_stressor,
reg_sine, reg_asym, or lm.}

\item{data}{A data frame object that contains all the entries to be cross-
validated on.}

\item{t_groups}{A vector of how the data entries are assigned to each group.}
}
\value{
If the object is of class mlm_stressor then a data frame will be
  returned otherwise a vector of the predictions will be returned.
}
\description{
This is core of cross validation both standard and using k-mean groups. This
 method is called by other cv methods of classes.
}
