% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_core.R
\name{cv_core}
\alias{cv_core}
\title{Cross Validation Function}
\usage{
cv_core(object, data, t_groups, ...)
}
\arguments{
\item{object}{Currently `"reg_sine", "reg_asym", "lm", "mlm_stressor"`
objects are accepted.}

\item{data}{A data frame object that has the same formula that was fitted on
the data.}

\item{t_groups}{The groups for cross validation: standard cross validation,
LOO cross_validation, or spatial cross validation.}

\item{...}{Additional arguments that are passed to the predict function.}
}
\value{
Either a vector of predictions for `"reg_sine", "reg_asym", "lm"` and
  a data frame for `"mlm_stressor"`
}
\description{
The machinery to run cross validation in that it subsets the test and train
 set based on the groups it receives.
}
\examples{
 t_groups <- sample.int(10)
 lm_test <- data_gen_lm(10)
 lm <- lm(Y ~ ., lm_test)
 lm_cv <- cv_core(lm, lm_test, t_groups)
 lm_cv
}
