% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_lm.R
\name{data_gen_lm}
\alias{data_gen_lm}
\title{Data Generation for Linear Regression}
\usage{
data_gen_lm(n, weight_vec = rep(1, 5), y_int = 0, resp_sd = 1, ...)
}
\arguments{
\item{n}{The number of observations for each parameter.}

\item{weight_vec}{The parameter coefficients where each entry represents the
coefficients for the additive linear model.}

\item{y_int}{The y-intercept term of the additive model.}

\item{resp_sd}{The standard deviation of the epsilon term to be added for
noise.}
}
\value{
A dataframe object with the n rows and the response variable with
 the number of parameters being equal to the number of columns from the
 weight matrix.
}
\description{
Creates a synthetic dataset from various parameters that represent a simple
 example of an additive \eqn{y = mx}.
}
