% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_sine.R
\name{data_gen_sine}
\alias{data_gen_sine}
\title{Data Generation for Sinusoidal Regression}
\usage{
data_gen_sine(
  n,
  weight_mat = matrix(rnorm(15), nrow = 3, ncol = 5),
  y_int = 0,
  resp_sd = 1,
  ...
)
}
\arguments{
\item{n}{The number of observations for each parameter.}

\item{weight_mat}{The parameter coefficients where each column represents
the coefficients and is three rows as each additive equation contains three
parameters. Defaulted to be 15 random numbers from the normal distribution.}

\item{y_int}{The y-intercept term of the additive model.}

\item{resp_sd}{The standard deviation of the epsilon term to be added for
noise.}
}
\value{
A dataframe object with the n rows and the response variable with
 the number of parameters being equal to the number of columns from the
 weight matrix.
}
\description{
Creates a synthetic dataset from various parameters that represent a simple
 example of an additive \eqn{y = asin(b(x - c))}.
}
