% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_center.R
\name{dist_cent}
\alias{dist_cent}
\title{Distance to Center}
\usage{
dist_cent(formula, data)
}
\arguments{
\item{formula}{A formula object}

\item{data}{A data frame object}
}
\value{
A vector of distances from the center.
}
\description{
Calculates the distance from center of the X parameter space.
}
\examples{
  data <- data_gen_lm(10)
  dist <- dist_cent(Y ~ ., data)
  dist
}
