% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_init.R
\name{mlm_init}
\alias{mlm_init}
\title{Compare Machine Learning Models}
\usage{
mlm_init(
  formula,
  data,
  fit_models,
  n_models = 9999,
  classification = FALSE,
  ...
)
}
\arguments{
\item{formula}{The regression formula or classification formula}

\item{data}{A dataframe object that includes the test data}

\item{fit_models}{A character vector with all the possible Machine Learning
 regressors that are currently being fit, the user may specify a subset of
 them using a character vector.
 \tabular{rl}{
   ada \tab AdaBoost Regressor \cr
   et \tab Extra Trees Regressor \cr
   lightgbm \tab Light Gradient Boosting Machine\cr
   gbr \tab Gradient Boosting Regressor\cr
   lr \tab Linear Regression\cr
   rf \tab Random Forest Regressor\cr
   ridge \tab Ridge Regression\cr
   knn \tab K Neighbors Regressor\cr
   dt \tab Decision Tree Regressor\cr
   dummy \tab Dummy Regressor\cr
   lar \tab Least Angle Regression\cr
   br \tab Bayesian Ridge\cr
   huber \tab Huber Regressor\cr
  omp \tab Orthogonal Matching Pursuit\cr
   lasso \tab Lasso Regression\cr
   en \tab Elastic Net\cr
   llar \tab Lasso Least Angle Regression\cr
   par \tab Passive Aggressive Regressor
 }
If classification is set to TRUE, these models can be used depending on user
specified but these are the default values for classification:
 \tabular{rl}{
   ada \tab AdaBoost Classifier \cr
   et \tab Extra Trees Classifier \cr
   lightgbm \tab Light Gradient Boosting Machine\cr
   gbr \tab Gradient Boosting Classifier\cr
   lr \tab Logistic Regression\cr
   rf \tab Random Forest Classifier\cr
   ridge \tab Ridge Classifier\cr
   knn \tab K Neighbors Classifier\cr
   dt \tab Decision Tree Classifier\cr
   dummy \tab Dummy Classifier\cr
   svm \tab SVM - Linear Kernel\cr
   lda \tab Linear Discriminant Analysis\cr
   nb \tab Naive Bayes \cr
   qda \tab Quadratic Discriminant Analysis
 }}

\item{n_models}{A defaulted integer to return the maximum number of models}

\item{classification}{A boolean value tag to indicate if classification
methods should be used.}

\item{...}{Additional arguments passed to the setup function in pycaret}
}
\value{
A list object that contains all the fitted models and the CV
  predictive accuracy.
}
\description{
Through the pycaret module from python, this function fits many machine
  learning models simultaneously with little to no coding on the user
  part. The core function to fitting the initial models.
}
