% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_mlm.R
\name{refit_mlm}
\alias{refit_mlm}
\title{Refit Machine Learning Models}
\usage{
refit_mlm(models, train_data, test_data)
}
\arguments{
\item{models}{The returned object of the [compare_mlm()]}

\item{train_data}{A dataframe object used for refitting excludes the test
data}

\item{test_data}{A dataframe object used for predictions}
}
\value{
A named list object of result from pycaret's [predict_model()].
}
\description{
Refits models fitted in the [compare_mlm()], and returns the predictions.
}
