% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(predictions, observed)
}
\arguments{
\item{predictions}{A data frame or vector object that is the same number of
rows or length as the length of observed values.}

\item{observed}{A vector of the observed results}
}
\value{
A finite value or a data frame of methods and their rmse values
}
\description{
Calculates the root mean square error.
}
\examples{
  # Using vignette data
  data(boston)
  data(mlm_vignette_boston_cv)
  rmse_boston <- rmse(mlm_vignette_boston_cv, boston$cmedv)
  rmse_boston
}
