% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_regression.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squarred Error (RMSE)}
\usage{
rmse(predicted, observed)
}
\arguments{
\item{predicted}{A data.frame or vector object that is the same number of
rows or length as the length of observed values.}

\item{observed}{A vector of the observed results.}
}
\description{
A function to calculate the RMSE.
}
