% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thinning}
\alias{thinning}
\title{Thinning algorithm for Models with Predict function}
\usage{
thinning(
  model,
  data,
  max = 0.95,
  min = 0.05,
  iter = 0.05,
  classification = FALSE
)
}
\arguments{
\item{model}{a model that is currently of class type "reg_sine", "reg_asym",
"lm", or "mlm_stressor".}

\item{data}{a data frame with all the data}

\item{max}{a numeric value in (0, 1] and greater than `min`,
defaulted to .95.}

\item{min}{a numeric value in (0, 1) and less than `max`, defaulted to .05.}

\item{iter}{a numeric value to indicate the step size, defaulted to .05}

\item{classification}{A boolean value defaulted `FALSE`, used for
`mlm_classification`.}
}
\value{
a list object where the first element is the RMSE values at each
  iteration and the second element being the predictions.
}
\description{
Fits various train size and test sizes.
}
\examples{
\dontrun{
 lm_test <- data_gen_lm(20)
 create_virtualenv()
 mlm_lm <- mlm_regressor(Y ~ ., lm_test)
 thin <- thinning(mlm_lm, lm_test, max = .8, min = .7, iter = .05)
 thin
}
}
